CLI_PY_GUI/gui_terminal/requirements.txt
# Core Dependencies
PyQt6>=6.5.0
pywinpty>=2.0.0; platform_system == "Windows"
ptyprocess>=0.7.0; platform_system != "Windows"
psutil>=5.9.0

# Networking and WebSockets
websockets>=11.0.0
requests>=2.31.0
aiohttp>=3.8.0

# Data and Configuration
pydantic>=2.0.0
pyyaml>=6.0.0
redis>=4.5.0

# Security
cryptography>=41.0.0

# Logging and Monitoring
structlog>=23.0.0